(#002) Integrate Eleventy-powered static web page
The new web page enables users to browse and download available packages and
templates. It is designed in Figma and implemented with Eleventy to
ensure fast, relieble builds.

- Initialize npm project and stablish the base folder structure.
- Configure Eleventy as a static dite generator (SSG) for optimal
  performance
- Add a build script ´npm run build´ to produce production-ready assets
  for GitHub Pages.
- Organize SCSS and template files for clear user-template integration.

This commit sets up the foundation of our software of our software
lifecycle, transitioning from a initial Waterfall plan to Agile
development practices.

Signed-off-by: Joseph Gallegos <jaremgallegosp@gmail.com>